authorizer: Adds option to integrate cached backends
This change means we now accept the option for the caller to create a
backend cache on the fly when recieving an AuthRep request.

The default option still remains calling 3scale directly with the
request, however in the case where a caller wants caching, we try create
a cache for the provided apisonator url and carry out the request via
the cache.

Any errors to the point of actually using the cache to make the network
request will mean the request will go via the default client.